requirements.txt
truedata-ws
pandas
asyncio
python-telegram-bot
